move convertToUTC
