+++ packages/components/src/DateTimePickers/DateTime/datetime-extraction.js
@@ -20,6 +20,22 @@ function isEmpty(value) {
 }
 
 /**
+ * Convert a date in local TZ to UTC
+ */
+function convertToUTC(date) {
+	return new Date(
+		Date.UTC(
+			date.getFullYear(),
+			date.getMonth(),
+			date.getDate(),
+			date.getHours(),
+			date.getMinutes(),
+			date.getSeconds(),
+		),
+	);
+}
+
+/**
  * Extract time
  * @param date {Date} The date to extract
  * @param useSeconds {boolean} Indicates if we should extract seconds
@@ -257,6 +273,7 @@ function updateDatetimeOnTimeChange(timePickerPayload, date, options) {
 }
 
 export {
+	convertToUTC,
 	extractParts,
 	extractPartsFromDateTime,
 	extractPartsFromTextInput,
